fix: add close method to stream iterator for connection release (#630)
* fix: add close method to stream iterator for connection release

* fix: add close method to stream iterator for connection release

* fix: add close method to stream iterator for connection release

* fix: add close method to stream iterator for connection release

* fix: add close method to stream iterator for connection release

* fix: add close method to stream iterator for connection release
