+++ java/src/main/java/com/baidubce/qianfan/core/builder/ChatBuilder.java
@@ -17,9 +17,9 @@
 package com.baidubce.qianfan.core.builder;
 
 import com.baidubce.qianfan.Qianfan;
+import com.baidubce.qianfan.core.StreamIterator;
 import com.baidubce.qianfan.model.chat.*;
 
-import java.util.Iterator;
 import java.util.List;
 
 public class ChatBuilder extends BaseBuilder<ChatBuilder> {
@@ -175,7 +175,7 @@ public class ChatBuilder extends BaseBuilder<ChatBuilder> {
         return super.getQianfan().chatCompletion(build());
     }
 
-    public Iterator<ChatResponse> executeStream() {
+    public StreamIterator<ChatResponse> executeStream() {
         return super.getQianfan().chatCompletionStream(build());
     }
 }
