+++ java/src/main/java/com/baidubce/qianfan/core/builder/CompletionBuilder.java
@@ -17,10 +17,10 @@
 package com.baidubce.qianfan.core.builder;
 
 import com.baidubce.qianfan.Qianfan;
+import com.baidubce.qianfan.core.StreamIterator;
 import com.baidubce.qianfan.model.completion.CompletionRequest;
 import com.baidubce.qianfan.model.completion.CompletionResponse;
 
-import java.util.Iterator;
 import java.util.List;
 
 public class CompletionBuilder extends BaseBuilder<CompletionBuilder> {
@@ -92,7 +92,7 @@ public class CompletionBuilder extends BaseBuilder<CompletionBuilder> {
         return super.getQianfan().completion(build());
     }
 
-    public Iterator<CompletionResponse> executeStream() {
+    public StreamIterator<CompletionResponse> executeStream() {
         return super.getQianfan().completionStream(build());
     }
 }
