+++ java/src/main/java/com/baidubce/qianfan/core/builder/Image2TextBuilder.java
@@ -17,10 +17,10 @@
 package com.baidubce.qianfan.core.builder;
 
 import com.baidubce.qianfan.Qianfan;
+import com.baidubce.qianfan.core.StreamIterator;
 import com.baidubce.qianfan.model.image.Image2TextRequest;
 import com.baidubce.qianfan.model.image.Image2TextResponse;
 
-import java.util.Iterator;
 import java.util.List;
 
 public class Image2TextBuilder extends BaseBuilder<Image2TextBuilder> {
@@ -140,7 +140,7 @@ public class Image2TextBuilder extends BaseBuilder<Image2TextBuilder> {
         return super.getQianfan().image2Text(build());
     }
 
-    public Iterator<Image2TextResponse> executeStream() {
+    public StreamIterator<Image2TextResponse> executeStream() {
         return super.getQianfan().image2TextStream(build());
     }
 }
