+++ java/src/main/java/com/baidubce/qianfan/core/builder/PluginBuilder.java
@@ -17,12 +17,12 @@
 package com.baidubce.qianfan.core.builder;
 
 import com.baidubce.qianfan.Qianfan;
+import com.baidubce.qianfan.core.StreamIterator;
 import com.baidubce.qianfan.model.plugin.PluginHistory;
 import com.baidubce.qianfan.model.plugin.PluginLLM;
 import com.baidubce.qianfan.model.plugin.PluginRequest;
 import com.baidubce.qianfan.model.plugin.PluginResponse;
 
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
@@ -102,7 +102,7 @@ public class PluginBuilder extends BaseBuilder<PluginBuilder> {
         return super.getQianfan().plugin(build());
     }
 
-    public Iterator<PluginResponse> executeStream() {
+    public StreamIterator<PluginResponse> executeStream() {
         return super.getQianfan().pluginStream(build());
     }
 }
