+++ java/src/main/java/com/baidubce/qianfan/util/http/HttpRequest.java
@@ -28,7 +28,6 @@ import org.apache.hc.core5.http.io.entity.StringEntity;
 import java.io.IOException;
 import java.lang.reflect.Type;
 import java.net.URI;
-import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
@@ -120,7 +119,7 @@ public class HttpRequest {
         return HttpClient.executeString(toClassicHttpRequest());
     }
 
-    public HttpResponse<Iterator<String>> executeSSE() throws IOException {
+    public HttpResponse<SSEIterator> executeSSE() throws IOException {
         return HttpClient.executeSSE(toClassicHttpRequest());
     }
